+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/procs/compare/v0.9.1...Unreleased) - ReleaseDate
 
+* [Fixed] Ctrl-C is ignored on Windows [#35](https://github.com/dalance/procs/issues/35)
+
 ## [v0.9.1](https://github.com/dalance/procs/compare/v0.9.0...v0.9.1) - 2020-01-24
 
 * [Fixed] clear screen at entering watch mode